+++ source/ftrack_connect_nuke/connector/nukeassets.py
@@ -337,7 +337,7 @@ class GizmoAsset(GenericAsset):
 
     def importAsset(self, iAObj=None):
         if iAObj.filePath.endswith('gizmo'):
-            resultingNode = nuke.createNode(iAObj.filePath)
+            resultingNode = nuke.createNode(iAObj.filePath.replace('\\', '/'))
             resultingNode['name'].setValue(iAObj.assetName)
             self.addFTab(resultingNode)
             self.setFTab(resultingNode, iAObj)
@@ -346,11 +346,11 @@ class GizmoAsset(GenericAsset):
     def changeVersion(self, iAObj=None, applicationObject=None):
 
         old_gizmo = nuke.toNode(applicationObject)
-
-        gizmo_path = os.path.dirname(iAObj.filePath)
+        file_path = iAObj.filePath.replace('\\', '/')
+        gizmo_path = os.path.dirname(file_path)
         nuke.pluginAddPath(gizmo_path)
 
-        new_gizmo = nuke.createNode(iAObj.filePath)
+        new_gizmo = nuke.createNode(file_path)
 
         # connect inputs
         for i in range(old_gizmo.inputs()):