+++ resource/hook/ftrack_connect_nuke_hook.py
@@ -8,7 +8,7 @@ import logging
 import re
 import os
 
-import ftrack
+import ftrack_api
 import ftrack_connect.application
 import ftrack_connect_nuke
 
@@ -18,7 +18,7 @@ class LaunchApplicationAction(object):
 
     identifier = 'ftrack-connect-launch-nuke'
 
-    def __init__(self, application_store, launcher):
+    def __init__(self, application_store, launcher, session):
         '''Initialise action with *applicationStore* and *launcher*.
 
         *applicationStore* should be an instance of
@@ -36,6 +36,7 @@ class LaunchApplicationAction(object):
 
         self.application_store = application_store
         self.launcher = launcher
+        self.session = session
 
     def is_valid_selection(self, selection):
         '''Return true if the selection is valid.'''
@@ -46,23 +47,28 @@ class LaunchApplicationAction(object):
             return False
 
         entity = selection[0]
-        task = ftrack.Task(entity['entityId'])
 
-        if task.getObjectType() != 'Task':
+        task = self.session.get(
+            'Task', entity['entityId']
+        )
+
+
+        if task is None:
             return False
 
         return True
 
     def register(self):
         '''Register discover actions on logged in user.'''
-        ftrack.EVENT_HUB.subscribe(
+        self.session.event_hub.subscribe(
             'topic=ftrack.action.discover and source.user.username={0}'.format(
                 getpass.getuser()
             ),
-            self.discover
+            self.discover,
+            priority=10
         )
 
-        ftrack.EVENT_HUB.subscribe(
+        self.session.event_hub.subscribe(
             'topic=ftrack.action.launch and source.user.username={0} '
             'and data.actionIdentifier={1}'.format(
                 getpass.getuser(), self.identifier
@@ -70,7 +76,7 @@ class LaunchApplicationAction(object):
             self.launch
         )
 
-        ftrack.EVENT_HUB.subscribe(
+        self.session.event_hub.subscribe(
             'topic=ftrack.connect.plugin.debug-information',
             self.get_version_information
         )
@@ -247,11 +253,12 @@ class ApplicationStore(ftrack_connect.application.ApplicationStore):
 class ApplicationLauncher(ftrack_connect.application.ApplicationLauncher):
     '''Custom launcher to modify environment before launch.'''
 
-    def __init__(self, application_store, plugin_path):
+    def __init__(self, application_store, plugin_path, session):
         '''.'''
         super(ApplicationLauncher, self).__init__(application_store)
 
         self.plugin_path = plugin_path
+        self.session = session
 
     def _getApplicationEnvironment(
         self, application, context=None
@@ -265,20 +272,32 @@ class ApplicationLauncher(ftrack_connect.application.ApplicationLauncher):
         )._getApplicationEnvironment(application, context)
 
         entity = context['selection'][0]
-        task = ftrack.Task(entity['entityId'])
-        taskParent = task.getParent()
+
+        task = self.session.get(
+            'Task', entity['entityId']
+        )
+
+        taskParent = task.get(
+            'parent'
+        )
 
         try:
-            environment['FS'] = str(int(taskParent.getFrameStart()))
+            environment['FS'] = str(
+                int(taskParent['custom_attributes'].get('fstart'))
+            )
+
         except Exception:
             environment['FS'] = '1'
 
         try:
-            environment['FE'] = str(int(taskParent.getFrameEnd()))
+            environment['FE'] = str(
+                int(taskParent['custom_attributes'].get('fend'))
+            )
+
         except Exception:
             environment['FE'] = '1'
 
-        environment['FTRACK_TASKID'] = task.getId()
+        environment['FTRACK_TASKID'] = task.get('id')
         environment['FTRACK_SHOTID'] = task.get('parent_id')
 
         nuke_plugin_path = os.path.abspath(
@@ -318,18 +337,20 @@ class ApplicationLauncher(ftrack_connect.application.ApplicationLauncher):
         return environment
 
 
-def register(registry, **kw):
+def register(session, **kw):
     '''Register hooks.'''
 
     logger = logging.getLogger(
         'ftrack_plugin:ftrack_connect_nuke_hook.register'
     )
 
-    # Validate that registry is the event handler registry. If not,
-    # assume that register is being called to regiter Locations or from a new
-    # or incompatible API, and return without doing anything.
-    if registry is not ftrack.EVENT_HANDLERS:
+
+    # Validate that session is an instance of ftrack_api.Session. If not,
+    # assume that register is being called from an old or incompatible API and
+    # return without doing anything.
+    if not isinstance(session, ftrack_api.session.Session):
         return
+
 
     # Create store containing applications.
     application_store = ApplicationStore()
@@ -343,9 +364,12 @@ def register(registry, **kw):
                     os.path.dirname(__file__), '..', 'ftrack_connect_nuke'
                 )
             )
-        )
+        ),
+        session=session
     )
 
     # Create action and register to respond to discover and launch actions.
-    action = LaunchApplicationAction(application_store, launcher)
+    action = LaunchApplicationAction(application_store, launcher, session)
     action.register()
+
+