+++ doc/release/release_notes.rst
@@ -11,10 +11,14 @@ Release notes
 .. release:: Upcoming
 
     .. change:: fixed
-        :tags: Action
+        :tags: API
+
+        Connector does not get initialized correctly.
+
+    .. change:: fixed
+        :tags: API
 
         NukeX does not get properly discovered under OSx.
-
 
 .. release:: 0.1.13
     :date: 2017-01-04