+++ doc/release/release_notes.rst
@@ -8,6 +8,15 @@
 Release notes
 *************
 
+.. release:: 1.0.1
+
+  :date: 2017-07-11
+
+  .. change:: fixed
+      :tags: Actions
+
+      Nuke is discovered under NukeStudio in Connect.
+
 .. release:: 1.0.0
   :date: 2017-07-07